pyproject.toml
[project]
name = "nadobro-bot"
version = "0.2.0"
description = "Nadobro — Telegram trading bot for Perps on Nado (Ink L2 CLOB DEX)"
requires-python = ">=3.11,<3.14"
dependencies = [
    "apscheduler==3.11.2",
    "cryptography==46.0.5",
    "eth-account>=0.8.0,<0.9.0",
    "nado-protocol==0.3.3",
    "openai==2.17.0",
    "pandas==2.2.3",
    "pinecone",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.2.1",
    "python-telegram-bot[webhooks]==22.6",
    "requests==2.32.5",
    "ta==0.11.0",
    "fastapi==0.115.0",
    "uvicorn[standard]==0.34.0",
    "websockets>=10.0.0,<14.0.0",
]
